+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [Unreleased]
 ### Added
+- Support of DDNet front, switch and tele layer
+- Support of Ninslash maps
+- Support of FNG2 maps
 - Object zone layers
 - Option to create a ZIP file that contains the current package and all its dependencies
 - Button to apply dilate and border_add to images
-- Support of DDNet front, switch and tele layer
-- Support of Ninslash maps
 - Each quad get its own global color in addition of the color set for each vertex
 
 ## Changed
@@ -16,6 +17,7 @@
 - Default packages are now opened in read-only mode by default
 
 ### Fixed
+- OpenFNG map import
 - Zoom-out can freeze the editor
 - Flickering cursor on some environments